Cleanup and preparation for inside-wrapper TPM refactoring.
Add recovery reason for already in recovery and need to reboot to
recovery to let the TPM init.

Add vboot_struct fields.

Fix type for keyblock flags param to SetTPMBootModeState().

BUG=none
TEST=make && make runtests

Reviewed-on: http://gerrit.chromium.org/gerrit/4599
Tested-by: Randall Spangler <rspangler@chromium.org>

+++ firmware/lib/include/tpm_bootmode.h
@@ -20,6 +20,6 @@
  */
 
 uint32_t SetTPMBootModeState(int developer_mode, int recovery_mode,
-                             int fw_keyblock_flags);
+                             uint64_t fw_keyblock_flags);
 
 #endif  /* VBOOT_REFERENCE_TPM_BOOTMODE_H_ */
